requirements.txt
pika<=1.3.2
SpeechRecognition<=3.1.2
openai<=0.28.1
sentencepiece<=0.1.99
sacremoses<=0.0.53
tensorflow<=2.12.0
numpy<=1.25.0
matplotlib==3.7.1
scikit-learn==1.2.2
opencv-contrib-python<=4.7.0
pandas<=2.1.1
keras<=2.12
librosa<=0.10.0.post2
python-multipart<=0.0.6
fastapi<=0.68.2
moviepy<=1.0.3
transformers==4.34.0
python-dotenv<=0.16.0
translate<=3.6.1
uvicorn<=0.24.0.post1
pydantic<=1.10.13
langchain